Refactor - tsdoc doesn't handle get fields, so I've gotta work around it.
